backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
google-generativeai==0.3.2
python-dotenv==1.0.1
click==8.1.7
requests==2.31.0
rich==13.7.0 